renamed package ugentstudent to student
